refactor: boardDetail api 분리
